Add files via upload
This is a very small 'web' software of sorts that allows you to read and listen to audiobooks in the same place. Given you have the audio books in a pdf file format, along side the mp3 audio for the audiobook. Nothing fancy just something I came up with thinking back on how I struggled with reading, so if this might help you in some way that'll be enough for me. Forgive me if this upload is wrong in any way this is my first upload Thank you!